api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["paragpurandare <paragpurandare50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
pydantic = "^2.11.7"
langgraph = "^0.5.1"
langchain = "^0.3.26"
langchain-google-genai = "^2.1.6"
langchain-community = "^0.3.27"
langchain-chroma = "^0.2.4"
pydantic-ai = "^0.3.6"
pygithub = "^2.6.1"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
python-dotenv = "^1.1.1"
docx2txt = "^0.9"
pypdf2 = "^3.0.1"
playwright = "^1.53.0"
fastembed = "^0.7.1"
protobuf = ">=6.30.0,<7.0.0"
ipython = "^9.4.0"
graphviz = "^0.21"
pypdf = "^5.7.0"
chromadb = "^1.0.15"
sentence-transformers = "^5.0.0"
supabase = ">=2.16.0,<3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"